abstract awaying the shared constructor so that phpcpd is happy about less duplicated code
